Move session feedback into shared directory
We can re-use the check answers page in the PP journey, alongside some
other bits of this. I'd quite like us to re-structure the directories in
this project based on section of the journey e.g. session delivery,
intervention progress, rather than everything in one big SP/PP-focused
directory.

This is a starting point, and if we think this is a good route to go
down, we can continue refactoring this.
