+++ server/routes/serviceProviderReferrals/serviceProviderReferralsController.ts
@@ -29,16 +29,16 @@ import ActionPlanNumberOfSessionsForm from './actionPlanNumberOfSessionsForm'
 import EditSessionPresenter from './editSessionPresenter'
 import EditSessionView from './editSessionView'
 import EditSessionForm from './editSessionForm'
-import PostSessionAttendanceFeedbackView from './postSessionAttendanceFeedbackView'
-import PostSessionAttendanceFeedbackPresenter from './postSessionAttendanceFeedbackPresenter'
-import PostSessionAttendanceFeedbackForm from './postSessionAttendanceFeedbackForm'
-import PostSessionFeedbackConfirmationPresenter from './postSessionFeedbackConfirmationPresenter'
-import PostSessionFeedbackConfirmationView from './postSessionFeedbackConfirmationView'
-import PostSessionBehaviourFeedbackPresenter from './postSessionBehaviourFeedbackPresenter'
-import PostSessionBehaviourFeedbackView from './postSessionBehaviourFeedbackView'
-import PostSessionBehaviourFeedbackForm from './postSessionBehaviourFeedbackForm'
-import PostSessionFeedbackCheckAnswersView from './postSessionFeedbackCheckAnswersView'
-import PostSessionFeedbackCheckAnswersPresenter from './postSessionFeedbackCheckAnswersPresenter'
+import PostSessionAttendanceFeedbackForm from '../sessionDelivery/postSessionFeedback/postSessionAttendanceFeedbackForm'
+import PostSessionAttendanceFeedbackPresenter from '../sessionDelivery/postSessionFeedback/postSessionAttendanceFeedbackPresenter'
+import PostSessionAttendanceFeedbackView from '../sessionDelivery/postSessionFeedback/postSessionAttendanceFeedbackView'
+import PostSessionBehaviourFeedbackForm from '../sessionDelivery/postSessionFeedback/postSessionBehaviourFeedbackForm'
+import PostSessionBehaviourFeedbackPresenter from '../sessionDelivery/postSessionFeedback/postSessionBehaviourFeedbackPresenter'
+import PostSessionBehaviourFeedbackView from '../sessionDelivery/postSessionFeedback/postSessionBehaviourFeedbackView'
+import PostSessionFeedbackCheckAnswersPresenter from '../sessionDelivery/postSessionFeedback/postSessionFeedbackCheckAnswersPresenter'
+import PostSessionFeedbackCheckAnswersView from '../sessionDelivery/postSessionFeedback/postSessionFeedbackCheckAnswersView'
+import PostSessionFeedbackConfirmationPresenter from '../sessionDelivery/postSessionFeedback/postSessionFeedbackConfirmationPresenter'
+import PostSessionFeedbackConfirmationView from '../sessionDelivery/postSessionFeedback/postSessionFeedbackConfirmationView'
 
 export default class ServiceProviderReferralsController {
   constructor(
