+++ server/routes/sessionDelivery/postSessionFeedback/postSessionBehaviourFeedbackForm.test.ts
@@ -1,4 +1,4 @@
-import TestUtils from '../../../testutils/testUtils'
+import TestUtils from '../../../../testutils/testUtils'
 import PostSessionBehaviourFeedbackForm from './postSessionBehaviourFeedbackForm'
 
 describe(PostSessionBehaviourFeedbackForm, () => {
