+++ server/views/sessionDelivery/postSessionFeedback/postSessionAttendanceFeedback.njk
@@ -3,14 +3,16 @@
 {% from "govuk/components/summary-list/macro.njk" import govukSummaryList %}
 {% from "govuk/components/textarea/macro.njk" import govukTextarea %}
 
-{% extends "./actionPlan/actionPlanFormTemplate.njk" %}
+{% extends "../../serviceProviderReferrals/actionPlan/actionPlanFormTemplate.njk" %}
 
 {% block formSection %}
+  {{ govukSummaryList(summaryListArgs) }}
+
   <form method="post">
     <input type="hidden" name="_csrf" value="{{ csrfToken }}">
-    {{ govukTextarea(textAreaArgs) }}
-
     {{ govukRadios(radioButtonArgs) }}
+
+    {{ govukTextarea(textAreaArgs) }}
 
     {{ govukButton({ text: "Save and continue" }) }}
   </form>
