+++ server/views/sessionDelivery/postSessionFeedback/postSessionBehaviourFeedback.njk
@@ -3,16 +3,14 @@
 {% from "govuk/components/summary-list/macro.njk" import govukSummaryList %}
 {% from "govuk/components/textarea/macro.njk" import govukTextarea %}
 
-{% extends "./actionPlan/actionPlanFormTemplate.njk" %}
+{% extends "../../serviceProviderReferrals/actionPlan/actionPlanFormTemplate.njk" %}
 
 {% block formSection %}
-  {{ govukSummaryList(summaryListArgs) }}
-
   <form method="post">
     <input type="hidden" name="_csrf" value="{{ csrfToken }}">
-    {{ govukRadios(radioButtonArgs) }}
-
     {{ govukTextarea(textAreaArgs) }}
+
+    {{ govukRadios(radioButtonArgs) }}
 
     {{ govukButton({ text: "Save and continue" }) }}
   </form>
