+++ server/views/sessionDelivery/postSessionFeedback/postSessionFeedbackConfirmation.njk
@@ -1,7 +1,7 @@
 {% from "govuk/components/summary-list/macro.njk" import govukSummaryList %}
 {% from "govuk/components/panel/macro.njk" import govukPanel %}
 
-{% extends "../partials/layout.njk" %}
+{% extends "../../partials/layout.njk" %}
 
 {% block pageTitle %}
   HMPPS Inteventions - GOV.UK
